+++ libvec/src/v2_set.c
@@ -6,13 +6,13 @@
 /*   By: cschoen <cschoen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 14:27:07 by cschoen           #+#    #+#             */
-/*   Updated: 2019/10/13 14:34:29 by cschoen          ###   ########.fr       */
+/*   Updated: 2020/02/15 20:49:15 by cschoen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libvec.h"
 
-t_vec2	v2_set(double u, double v)
+t_vec2	v2_set(float u, float v)
 {
 	t_vec2	new_v2;
