+++ libvec/src/v3_length.c
@@ -6,14 +6,14 @@
 /*   By: cschoen <cschoen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 03:33:21 by cschoen           #+#    #+#             */
-/*   Updated: 2019/10/07 03:36:06 by cschoen          ###   ########.fr       */
+/*   Updated: 2020/02/15 20:49:15 by cschoen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libvec.h"
 #include <math.h>
 
-double	v3_length(t_vec3 v)
+float	v3_length(t_vec3 v)
 {
 	return (sqrt(v.x * v.x + v.y * v.y + v.z * v.z));
 }
