+++ libvec/src/v3_rot_x.c
@@ -6,13 +6,13 @@
 /*   By: cschoen <cschoen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 21:17:48 by cschoen           #+#    #+#             */
-/*   Updated: 2019/10/14 05:39:34 by cschoen          ###   ########.fr       */
+/*   Updated: 2020/02/15 20:49:15 by cschoen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libvec.h"
 
-t_vec3	v3_rot_x(t_vec3 v, double angle)
+t_vec3	v3_rot_x(t_vec3 v, float angle)
 {
 	t_vec3	new_v;
